# lib/api-zod/src/generated/types/activityBucket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spotify Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityBucket {
  label: string;
  count: number;
}