# lib/api-zod/src/generated/types/activityData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spotify Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityBucket } from './activityBucket';
import type { TimelinePoint } from './timelinePoint';

export interface ActivityData {
  byHour: ActivityBucket[];
  byDayOfWeek: ActivityBucket[];
  byMonth: ActivityBucket[];
  timeline: TimelinePoint[];
}