# lib/api-zod/src/generated/types/artistDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spotify Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { SpotifyAlbum } from './spotifyAlbum';
import type { SpotifyArtist } from './spotifyArtist';
import type { SpotifyTrack } from './spotifyTrack';

export interface ArtistDetail {
  artist: SpotifyArtist;
  topTracks: SpotifyTrack[];
  relatedArtists: SpotifyArtist[];
  albums: SpotifyAlbum[];
  /** @nullable */
  userRank?: number | null;
}