# lib/api-zod/src/generated/types/savedAlbumsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spotify Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { SpotifyAlbum } from './spotifyAlbum';

export interface SavedAlbumsResponse {
  total: number;
  items: SpotifyAlbum[];
}