# lib/api-zod/src/generated/types/savedTracksResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spotify Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { SpotifyTrack } from './spotifyTrack';

export interface SavedTracksResponse {
  total: number;
  items: SpotifyTrack[];
}